requirements.txt
streamlit
py3dmol
rdkit-pypi
